Use single callback endpoint
It's a bit of a pain to have two separate callback paths in omniauth.
There is a callback_path option, and it does take a block, but I'm a
little unclear about how it knows how to start the normal OAuth flow
if you use a block for this option.

To keep things simple I went with the single callback endpoint.