Clear stale session IDs in callback
When the OAuth callback is triggered, there may already be session IDs
in the session, and sometimes those IDs will be for a different shop.
This can happen if the user was previously accessing the app for shop 1
and now accesses the app for shop 2.

These edge cases can result in incorrectly attempting to use a token on
the wrong shop, which fails.

To avoid this scenario, we check that the IDs in the session belong to
the shop that triggered the callback, and clear the values if not.